Initial Commit
The base functionality is in place, and the emulator builds
and runs, however the bitwidth is not working correctly so
it only functions in 8 bit mode, regardless of the width set.
This suggests that my c-style bithacking won't work here, and
that I'll need to implement something more proper but slower.